requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.12.0
django-redis==5.2.0
djangorestframework==3.13.1
Faker==13.12.0
mysqlclient==2.1.0
packaging==21.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
wrapt==1.14.1
